+++ src/Exception/TranslatableAttributeException.php
@@ -5,18 +5,8 @@ namespace Jtl\Connector\Core\Exception;
 class TranslatableAttributeException extends \Exception
 {
     public const
-        TYPE_UNKNOWN = 10,
         VALUE_TYPE_INVALID = 20,
         DECODING_VALUE_FAILED = 30;
-
-    /**
-     * @param string $type
-     * @return TranslatableAttributeException
-     */
-    public static function typeUnknown(string $type): self
-    {
-        return new self(sprintf('"%s" is not a valid translatable attribute data type', $type), self::TYPE_UNKNOWN);
-    }
 
     /**
      * @param string $attributeName